SECT-MOBILE-IOS-DUREE: epreuve.duree est Int32, pas besoin de .intValue
Dernière erreur : 'value of type Int32 has no member intValue'
epreuve.duree est déjà un Int32 (pas KotlinInt), donc Int(epreuve.duree)
suffit sans .intValue.

+++ mobile/iosApp/iosApp/ViewModels/PassationViewModel.swift
@@ -48,7 +48,7 @@ class PassationViewModel: ObservableObject {
             if let tempsRestant = session.tempsRestant, Int(tempsRestant.intValue) > 0 {
                 totalSeconds = Int(tempsRestant.intValue)
             } else if let epreuve = self.epreuve {
-                totalSeconds = Int(epreuve.duree.intValue) * 60
+                totalSeconds = Int(epreuve.duree) * 60
             } else {
                 totalSeconds = 0
             }
